+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,24 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CardsGame">
+        <activity android:name=".MainActivity"
+            android:screenOrientation="landscape"
+            >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+
+        </activity>
+
+
+
         <activity android:name=".WinnerActivity"
             android:screenOrientation="landscape"
             ></activity>
-        <activity android:name=".MainActivity"
+        <activity android:name=".GameActivity"
         android:screenOrientation="landscape"
             android:configChanges="layoutDirection|orientation|screenLayout|screenSize">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
     </application>
