+++ app/src/main/res/layout/activity_game.xml
@@ -1,0 +1,121 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".GameActivity"
+    android:orientation="horizontal"
+    android:background="#00695C"
+    tools:ignore="ExtraText">
+    <ProgressBar
+        android:id="@+id/Game_progressBar"
+        style="@android:style/Widget.ProgressBar.Horizontal"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/main_BTN_start"
+        android:layout_centerHorizontal="true"
+        />
+
+
+    <Button
+        android:id="@+id/main_BTN_start"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Start Game"
+        android:textColor="@color/black"
+        android:visibility="visible"
+        android:layout_centerInParent="true"/>
+
+    <Button
+        android:visibility="invisible"
+        android:id="@+id/main_BTN_drawCard"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="100dp"
+        android:textColor="@color/black"
+        android:text="Draw A Card"
+        android:layout_centerInParent="true" />
+
+    <LinearLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="200dp"
+        android:layout_height="match_parent"
+        android:paddingLeft="16dp"
+        android:paddingRight="16dp"
+        android:orientation="vertical"
+        android:gravity="center"
+        android:layout_toLeftOf="@id/main_BTN_start"
+        android:layout_marginRight="20dp">
+        <ImageView
+            android:id="@+id/main_IMG_woman"
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+            android:src="@drawable/woman"
+            android:layout_marginBottom="10dp"/>
+
+        <ImageView
+            android:id="@+id/main_IMG_unfolded1"
+            android:layout_width="200dp"
+            android:layout_height="200dp"
+            android:src="@drawable/img_unfolded" />
+
+        <TextView
+            android:id="@+id/player1_counter"
+            android:textSize="40sp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="0"
+            android:textColor="@color/black"
+            />
+    </LinearLayout>
+
+
+    <LinearLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="200dp"
+        android:layout_height="match_parent"
+        android:paddingLeft="16dp"
+        android:paddingRight="16dp"
+        android:orientation="vertical"
+        android:gravity="center"
+        android:layout_toRightOf="@id/main_BTN_start"
+        android:layout_marginLeft="20dp">
+        <ImageView
+            android:id="@+id/main_IMG_man"
+            android:layout_width="150dp"
+            android:layout_height="100dp"
+            android:src="@drawable/man"
+            android:layout_marginBottom="10dp"/>
+
+        <ImageView
+
+            android:id="@+id/main_IMG_unfolded2"
+            android:layout_width="200dp"
+            android:layout_height="200dp"
+            android:src="@drawable/img_unfolded" />
+
+        <TextView
+            android:id="@+id/player2_counter"
+            android:textSize="40dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="0"
+            android:textColor="@color/black"
+            />
+
+
+
+    </LinearLayout>
+
+
+
+
+
+
+
+
+
+
+</RelativeLayout>
