+++ app/src/main/res/layout/activity_main.xml
@@ -1,112 +1,80 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    xmlns:tools="http://schemas.android.com/apk/res-auto"
     android:orientation="horizontal"
-    android:background="#00695C"
-    tools:ignore="ExtraText">
-
-    <Button
-        android:id="@+id/main_BTN_start"
+    android:background="#00695C">
+    <TextView
+    android:id="@+id/Main_TXT_CardGame"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="CARD GAME"
+    android:layout_centerHorizontal="true"
+    android:textColor="@color/offWhite_ron"
+    android:textSize="35dp"
+    android:textStyle="bold"
+    android:shadowColor="@color/black"
+    android:shadowDx="2"
+    android:shadowDy="2"
+    android:shadowRadius="2"
+        />
+    <ImageView
+        android:id="@+id/Main_straight_flush"
+        android:src="@drawable/royal_straight_flush"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Start Game"
-        android:textColor="@color/black"
-        android:visibility="visible"
-        android:layout_centerInParent="true"/>
+        android:layout_centerInParent="true"
+        android:shadowColor="@color/black"
+        android:shadowDx="2"
+        android:shadowDy="2"
+        android:shadowRadius="1.9"
+        android:adjustViewBounds="true"
 
-    <Button
-        android:visibility="invisible"
-        android:id="@+id/main_BTN_drawCard"
+
+
+
+        />
+
+    <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="100dp"
-        android:textColor="@color/black"
-        android:text="Draw A Card"
-        android:layout_centerInParent="true" />
-
-    <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="200dp"
-        android:layout_height="match_parent"
+        android:gravity="bottom"
+        android:orientation="horizontal"
+        android:layout_centerHorizontal="true"
+        android:layout_alignParentBottom="true"
         android:paddingLeft="16dp"
-        android:paddingRight="16dp"
-        android:orientation="vertical"
-        android:gravity="center"
-        android:layout_toLeftOf="@id/main_BTN_start"
-        android:layout_marginRight="20dp">
-        <ImageView
-            android:id="@+id/main_IMG_woman"
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            android:src="@drawable/woman"
-            android:layout_marginBottom="10dp"/>
+        android:paddingRight="16dp">
 
-        <ImageView
-            android:id="@+id/main_IMG_unfolded1"
-            android:layout_width="200dp"
-            android:layout_height="200dp"
-            android:src="@drawable/img_unfolded" />
-
-        <TextView
-            android:id="@+id/player1_counter"
-            android:textSize="40sp"
+        <Button
+            android:id="@+id/main_BTN_cardGame"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="0"
+            android:text="Card Game"
             android:textColor="@color/black"
-            />
-    </LinearLayout>
+            android:visibility="visible" />
 
-
-    <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="200dp"
-        android:layout_height="match_parent"
-        android:paddingLeft="16dp"
-        android:paddingRight="16dp"
-        android:orientation="vertical"
-        android:gravity="center"
-        android:layout_toRightOf="@id/main_BTN_start"
-        android:layout_marginLeft="20dp">
-        <ImageView
-            android:id="@+id/main_IMG_man"
-            android:layout_width="150dp"
-            android:layout_height="100dp"
-            android:src="@drawable/man"
-            android:layout_marginBottom="10dp"/>
-
-        <ImageView
-
-            android:id="@+id/main_IMG_unfolded2"
-            android:layout_width="200dp"
-            android:layout_height="200dp"
-            android:src="@drawable/img_unfolded" />
-
-        <TextView
-            android:id="@+id/player2_counter"
-            android:textSize="40dp"
+        <Button
+            android:id="@+id/main_BTN_Instructions"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="0"
+            android:layout_centerInParent="true"
+            android:text="Instructions"
             android:textColor="@color/black"
-            />
+            android:visibility="visible" />
 
+        <Button
+            android:id="@+id/main_BTN_HighScores"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerInParent="true"
+            android:text="HighScores"
+            android:textColor="@color/black"
+            android:visibility="visible" />
 
 
     </LinearLayout>
-
-
-
-
-
-
-
-
 
 
 </RelativeLayout>
